+++ WebApplication/Backend/src/main/java/com/its/gestionale/repository/UtenteRepository.java
@@ -9,7 +9,7 @@ import com.its.gestionale.entity.Ruolo;
 import com.its.gestionale.entity.Utente;
 
 @Repository
-public interface UtenteRepository extends JpaRepository<Utente, Long> {
+public interface UtenteRepository extends JpaRepository<Utente, Integer> {
 
     // Trova un utente per username — utile per login
     Optional<Utente> findByUsername(String username);
